requirements.txt
apache-airflow-providers-google
dbt-bigquery
pandas
google-cloud-bigquery
requests